requirements.txt
python-dotenv==1.0.0
requests==2.31.0
edge-tts>=7.0.0,<8.0.0
ffmpeg-python==0.2.0